requirements.txt
Flask==2.2.3
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
tensorflow>=2.16,<2.20
numpy>=1.25,<2.0
psycopg2-binary==2.9.6
werkzeug==2.3.4
Pillow==9.5.0
flask-cors
flask-jwt-extended
flask-migrate
flask-sqlalchemy
